FAERS class

Provide a container for FAERS Quarterly Data file

# S4 method for class 'FAERS'
show(object)

# S4 method for class 'FAERSascii'
show(object)

faers_data(object, ...)

# S4 method for class 'FAERS'
faers_data(object)

faers_year(object)

# S4 method for class 'FAERS'
faers_year(object)

faers_quarter(object)

# S4 method for class 'FAERS'
faers_quarter(object)

faers_period(object)

# S4 method for class 'FAERS'
faers_period(object)

faers_meddra(object, ...)

# S4 method for class 'FAERS'
faers_meddra(object, use = NULL)

faers_deleted_cases(object, ...)

# S4 method for class 'FAERSascii'
faers_deleted_cases(object)

faers_header(object)

# S4 method for class 'FAERSxml'
faers_header(object)

Arguments

object

A FAERS object.

...

Other arguments passed to specific methods.

use

A string, what meddra data to use, "hierarchy" or "smq". If NULL, a MedDRA will be returned. Only used when object has been standardized

Value

See details.

Details

faers_data: Extract the data slot.

faers_year: Extract the year slot.

faers_quarter: Extract the quarter slot.

faers_period: A data.table combine the year and quarter slot.

faers_meddra: Extract the meddra slot. If object have never been standardized, always return NULL.

faers_deleted_cases: Extract the deletedCases slot.

Slots

year

An integer specifies the year information.

quarter

A string specifies the quarter information.

data

For FAERSxml, a data.table. For FAERSascii, a list of data.table.

meddra

A MedDRA or NULL representing the meddra data used for standardization.

format

A string of "ascii" or "xml" indicates the file format used.

deletedCases

An atomic character, as of 2019 Quarter one there are new files that lists deleted cases. faers_dedup will remove cases in this slot.

standardization

A bool, indicates whether standardization has been performed.

deduplication

A bool, indicates whether deduplication has been performed.

Examples

# ususaly we use faers() function to create a `FAERS` object
# you must change `dir`, as the file included in the package is sampled
data <- faers(2004, "q1",
    dir = system.file("extdata", package = "faers"),
    compress_dir = tempdir()
)
#> Finding 1 file already downloaded: aers_ascii_2004q1.zip
faers_data(data)
#> $demo
#>       year quarter primaryid  caseid i_f_code foll_seq     image event_dt
#>      <int>  <char>    <char>  <char>   <char>    <int>    <char>    <int>
#>   1:  2004      q1   4263764 4060920        I       NA 4263764-6 20020101
#>   2:  2004      q1   4263927 4064250        I       NA 4263927-X       NA
#>   3:  2004      q1   4264001 4062524        I       NA 4264001-9 20031218
#>   4:  2004      q1   4264319 4064506        I       NA 4264319-X 20031216
#>   5:  2004      q1   4266745 4056689        I       NA 4266745-1 20030529
#>   6:  2004      q1   4268568 4059382        I       NA 4268568-6       NA
#>   7:  2004      q1   4268586 3994383        F       NA 4268586-8 20030826
#>   8:  2004      q1   4269553 4061181        I       NA 4269553-0 20040105
#>   9:  2004      q1   4269590 4062932        I       NA 4269590-6 20030602
#>  10:  2004      q1   4269894 4068631        I       NA 4269894-7 20030805
#>  11:  2004      q1   4270064 4062312        I       NA 4270064-7       NA
#>  12:  2004      q1   4271501 4007339        F        1 4271501-4 20020729
#>  13:  2004      q1   4272687 4072288        I       NA 4272687-8 20031220
#>  14:  2004      q1   4274829 4073121        I       NA 4274829-7       NA
#>  15:  2004      q1   4276604 4019165        F        2 4276604-6       NA
#>  16:  2004      q1   4276994 4062516        F        2 4276994-4 20031101
#>  17:  2004      q1   4277521 4027660        F        1 4277521-8       NA
#>  18:  2004      q1   4279336 4022360        F        1 4279336-3 20030911
#>  19:  2004      q1   4280316 4049927        F        1 4280316-2 20030617
#>  20:  2004      q1   4280654 4081346        I       NA 4280654-3 20040112
#>  21:  2004      q1   4281199 4093363        I       NA 4281199-7       NA
#>  22:  2004      q1   4281267 4074460        I       NA 4281267-X 20031111
#>  23:  2004      q1   4283791 4091687        I       NA 4283791-2       NA
#>  24:  2004      q1   4283960 4008465        F       NA 4283960-1 20010717
#>  25:  2004      q1   4284513 4084187        I       NA 4284513-1 20030101
#>  26:  2004      q1   4284876 4086360        I       NA 4284876-7       NA
#>  27:  2004      q1   4285379 4078728        F        1 4285379-6 20030813
#>  28:  2004      q1   4286029 4087281        I       NA 4286029-5       NA
#>  29:  2004      q1   4286712 4072824        F        1 4286712-1 20031209
#>  30:  2004      q1   4286757 4081521        F       NA 4286757-1 20031214
#>  31:  2004      q1   4286940 4088359        I       NA 4286940-5 20030121
#>  32:  2004      q1   4287019 4088362        I       NA 4287019-9 20040121
#>  33:  2004      q1   4289563 4091829        I       NA 4289563-7 20030901
#>        mfr_dt   fda_dt rept_cod                                  mfr_num
#>         <int>    <int>   <char>                                   <char>
#>   1: 20031219 20040102      EXP                             USA031255171
#>   2: 20031209 20040102      EXP                                B0317710A
#>   3: 20031219 20040102      EXP                    JP-JNJFOC-20031204393
#>   4: 20031218 20040105      EXP                             MEDI-0001221
#>   5: 20040105 20040108      EXP             FR-GLAXOSMITHKLINE-B0318977A
#>   6: 20040107 20040112      EXP                          PHBS2004JP00548
#>   7: 20030904 20040112      EXP             US-GLAXOSMITHKLINE-A0424697A
#>   8: 20040106 20040113      EXP US-BRISTOL-MYERS SQUIBB COMPANY-12471421
#>   9:       NA 20040112      DIR                                     <NA>
#>  10:       NA 20040112      DIR                                     <NA>
#>  11: 20030204 20040114      PER             US-GLAXOSMITHKLINE-A0395406A
#>  12: 20040106 20040113      EXP                            TPA2003A01222
#>  13: 20040109 20040115      EXP                               04-01-0044
#>  14: 20031231 20040115      EXP                             USA040155657
#>  15: 20040108 20040120      EXP                         US-SHR-03-015202
#>  16: 20040105 20040120      EXP                       03P-062-0245595-00
#>  17: 20040105 20040121      EXP                           DSA_23477_2003
#>  18: 20040112 20040122      EXP                    FR-JNJFOC-20030905591
#>  19: 20040115 20040126      EXP                                 UK059122
#>  20: 20040119 20040123      EXP                             200410213GDS
#>  21: 20030911 20040108      PER                         HQWYE745017SEP03
#>  22: 20031201 20040123      PER                                DL2003219
#>  23: 20030415 20040115      PER                         NSADSS2003017923
#>  24: 20031002 20040202      EXP                        WAES 0310USA00338
#>  25:       NA 20040130      DIR                                     <NA>
#>  26: 20040119 20040202      EXP                          A001-002-005955
#>  27: 20040121 20040202      EXP                             2004194398JP
#>  28: 20040122 20040203      EXP                               2004004976
#>  29: 20040126 20040204      EXP                             2004193432JP
#>  30: 20040127 20040206      EXP             IT-GLAXOSMITHKLINE-B0321668A
#>  31: 20040129 20040205      EXP                         USA-2004-0012715
#>  32: 20040121 20040205      EXP                            US_0401100140
#>  33: 20031020 20040126      PER                             USA030948470
#>                                    mfr_sndr   age age_cod gender  e_sub    wt
#>                                      <char> <num>  <char> <char> <char> <num>
#>   1:                  ELI LILLY AND COMPANY    68      YR      F      N  82.0
#>   2: GLAXOSMITHKLINE GLOBAL CLINICAL SAFETY    58      YR      F      N    NA
#>   3:                         CENTOCOR, INC.    53      YR      F      N  36.8
#>   4:                         MEDIMUNE, INC.    NA    <NA>      F      N    NA
#>   5:                        GLAXOSMITHKLINE    48      YR      F      Y    NA
#>   6:         NOVARTIS PHARMACEUTICALS CORP.    NA    <NA>      M      Y    NA
#>   7:                        GLAXOSMITHKLINE    13      YR      F      Y    NA
#>   8:           BRISTOL-MYERS SQUIBB COMPANY    10     DEC      F      Y    NA
#>   9:                                   <NA>    57      YR      M      N    NA
#>  10:                                   <NA>    90      YR      M      N    NA
#>  11:                        GLAXOSMITHKLINE    NA    <NA>   <NA>      Y    NA
#>  12:  TAKEDA PHARMACEUTICALS NORTH AMERICA.    71      YR      M      N 160.0
#>  13:             IVAX PHARMACEUTICALS, INC.    71      YR      F      N    NA
#>  14:                  ELI LILLY AND COMPANY    79      YR      F      N    NA
#>  15:                    BERLEX LABORATORIES    NA    <NA>      F      N  79.4
#>  16:                  PPD PHARMACOVIGILANCE    NA    <NA>      F      N    NA
#>  17:           BIOVAIL PHARMACEUTICALS INC.    61      YR      M      N 204.0
#>  18:                         CENTOCOR, INC.    31      YR      F      N    NA
#>  19:                                  AMGEN    65      YR      F      N  63.0
#>  20:      BAYER PHARMACEUTICALS CORPORATION    17      YR      F      N    NA
#>  21:                        WYETH LABS (RA)    NA    <NA>      F      N    NA
#>  22:                DANCO LABORATORIES, LLC    NA    <NA>      F      N    NA
#>  23:       JOHNSON + JOHNSON PHARMACEUTICAL    85      YR      F      N    NA
#>  24:            MERCK HUMAN HEALTH DIVISION    55      YR      M      Y    NA
#>  25:                                   <NA>    73      YR      M      N    NA
#>  26:                             EISAI INC.    NA    <NA>      F      N    NA
#>  27:                            PFIZER INC.    72      YR      F      N  52.0
#>  28:             WWS PFIZER PHARMACEUTICALS    NA    <NA>      M      N    NA
#>  29:                            PFIZER INC.    53      YR      M      N    NA
#>  30:                        GLAXOSMITHKLINE    17      YR      F      Y    NA
#>  31:                     PURDUE PHARMA L.P.    38      YR      F      N    NA
#>  32:                  ELI LILLY AND COMPANY    NA    <NA>      F      N    NA
#>  33:                  ELI LILLY AND COMPANY    79      YR      F      N    NA
#>      wt_cod  rept_dt occp_cod death_dt to_mfr confid    v23 caseversion
#>      <char>    <int>   <char>   <lgcl> <char> <char> <lgcl>       <int>
#>   1:     KG 20031223     <NA>       NA   <NA>   <NA>     NA           0
#>   2:   <NA> 20031219     <NA>       NA   <NA>   <NA>     NA           0
#>   3:     KG 20031231       MD       NA   <NA>   <NA>     NA           0
#>   4:   <NA> 20031231       MD       NA   <NA>   <NA>     NA           0
#>   5:   <NA> 20040108       CN       NA   <NA>   <NA>     NA           0
#>   6:   <NA> 20040109       MD       NA   <NA>   <NA>     NA           0
#>   7:   <NA> 20040112       CN       NA   <NA>   <NA>     NA           0
#>   8:   <NA> 20040113       OT       NA   <NA>   <NA>     NA           0
#>   9:   <NA>       NA       PH       NA      N      Y     NA           0
#>  10:   <NA>       NA     <NA>       NA      N      Y     NA           0
#>  11:   <NA> 20040114       CN       NA   <NA>   <NA>     NA           0
#>  12:    LBS 20040112       MD       NA   <NA>   <NA>     NA           0
#>  13:   <NA> 20040109       OT       NA   <NA>   <NA>     NA           0
#>  14:   <NA> 20040107     <NA>       NA   <NA>   <NA>     NA           0
#>  15:     KG 20040116       CN       NA   <NA>   <NA>     NA           0
#>  16:   <NA> 20040105       MD       NA   <NA>   <NA>     NA           0
#>  17:    LBS 20040120       MD       NA   <NA>   <NA>     NA           0
#>  18:   <NA> 20040121       MD       NA   <NA>   <NA>     NA           0
#>  19:     KG 20040123     <NA>       NA   <NA>   <NA>     NA           0
#>  20:   <NA> 20040122       MD       NA   <NA>   <NA>     NA           0
#>  21:   <NA> 20031223     <NA>       NA   <NA>   <NA>     NA           0
#>  22:   <NA> 20031229       OT       NA   <NA>   <NA>     NA           0
#>  23:   <NA> 20040108       OT       NA   <NA>   <NA>     NA           0
#>  24:   <NA> 20040203       MD       NA   <NA>   <NA>     NA           0
#>  25:   <NA> 20030124       OT       NA      Y      N     NA           0
#>  26:   <NA> 20040123       MD       NA   <NA>   <NA>     NA           0
#>  27:     KG 20040130       MD       NA   <NA>   <NA>     NA           0
#>  28:   <NA> 20040202       CN       NA   <NA>   <NA>     NA           0
#>  29:   <NA> 20040202       OT       NA   <NA>   <NA>     NA           0
#>  30:   <NA> 20040206       CN       NA   <NA>   <NA>     NA           0
#>  31:   <NA> 20040204       OT       NA   <NA>   <NA>     NA           0
#>  32:   <NA> 20040128     <NA>       NA   <NA>   <NA>     NA           0
#>  33:   <NA> 20031208       MD       NA   <NA>   <NA>     NA           0
#>      age_in_years country_code    sex
#>             <num>       <char> <char>
#>   1:           68         <NA> Female
#>   2:           58         <NA> Female
#>   3:           53         <NA> Female
#>   4:           NA         <NA> Female
#>   5:           48         <NA> Female
#>   6:           NA         <NA>   Male
#>   7:           13         <NA> Female
#>   8:          120         <NA> Female
#>   9:           57         <NA>   Male
#>  10:           90         <NA>   Male
#>  11:           NA         <NA>   <NA>
#>  12:           71         <NA>   Male
#>  13:           71         <NA> Female
#>  14:           79         <NA> Female
#>  15:           NA         <NA> Female
#>  16:           NA         <NA> Female
#>  17:           61         <NA>   Male
#>  18:           31         <NA> Female
#>  19:           65         <NA> Female
#>  20:           17         <NA> Female
#>  21:           NA         <NA> Female
#>  22:           NA         <NA> Female
#>  23:           85         <NA> Female
#>  24:           55         <NA>   Male
#>  25:           73         <NA>   Male
#>  26:           NA         <NA> Female
#>  27:           72         <NA> Female
#>  28:           NA         <NA>   Male
#>  29:           53         <NA>   Male
#>  30:           17         <NA> Female
#>  31:           38         <NA> Female
#>  32:           NA         <NA> Female
#>  33:           79         <NA> Female
#>  [ reached 'max' / getOption("max.print") -- omitted 68 rows ]
#> 
#> $drug
#>       year quarter primaryid   drug_seq role_cod  drugname val_vbm  route
#>      <int>  <char>    <char>      <int>   <char>    <char>   <int> <char>
#>   1:  2004      q1   4263764 1004492874       PS   HUMALOG       1   <NA>
#>   2:  2004      q1   4263764 1004515789       SS HUMULIN L       1   <NA>
#>   3:  2004      q1   4263764 1004515793       SS HUMULIN R       1   <NA>
#>   4:  2004      q1   4263927 1004493661       PS   AVANDIA       1   ORAL
#>   5:  2004      q1   4263927 1004529123        C GLYBURIDE       1   <NA>
#>  ---                                                                     
#> 337:  2004      q1   4373559 1004924622        C     ZETIA       1   <NA>
#> 338:  2004      q1   4373586 1004923625       PS   CRESTOR       1   ORAL
#> 339:  2004      q1   4373706 1004923791       PS   CRESTOR       1   <NA>
#> 340:  2004      q1   4373706 1004925107       SS    NIACIN       1   <NA>
#> 341:  2004      q1   4373706 1004925121        C   LIPITOR       1   <NA>
#>        dose_vbm dechal rechal lot_num exp_dt nda_num    v13
#>          <char> <char> <char>  <char>  <int>  <char> <lgcl>
#>   1:       <NA>      D   <NA>    <NA>     NA   20563     NA
#>   2: 15 U/2 DAY      D   <NA>    <NA>     NA    <NA>     NA
#>   3:       <NA>      N   <NA>    <NA>     NA    <NA>     NA
#>   4:       ORAL      Y      D    <NA>     NA   21071     NA
#>   5:       <NA>   <NA>   <NA>    <NA>     NA    <NA>     NA
#>  ---                                                       
#> 337:       <NA>   <NA>   <NA>    <NA>     NA    <NA>     NA
#> 338:   10 MG PO      D      D    <NA>     NA   21366     NA
#> 339:       <NA>      D      D    <NA>     NA   21366     NA
#> 340:       <NA>      D      D    <NA>     NA    <NA>     NA
#> 341:       <NA>   <NA>   <NA>    <NA>     NA    <NA>     NA
#> 
#> $indi
#>       year quarter primaryid indi_drug_seq
#>      <int>  <char>    <char>         <int>
#>   1:  2004      q1   4263927    1004493661
#>   2:  2004      q1   4264001    1004493811
#>   3:  2004      q1   4264001    1004520441
#>   4:  2004      q1   4264001    1004520538
#>   5:  2004      q1   4264319    1004494389
#>  ---                                      
#> 153:  2004      q1   4338901    1004786888
#> 154:  2004      q1   4346899    1004815291
#> 155:  2004      q1   4346899    1004815291
#> 156:  2004      q1   4347265    1004816501
#> 157:  2004      q1   4373559    1004923573
#>                                       indi_pt
#>                                        <char>
#>   1:  DIABETES MELLITUS NON-INSULIN-DEPENDENT
#>   2:                     RHEUMATOID ARTHRITIS
#>   3:                     RHEUMATOID ARTHRITIS
#>   4:                     RHEUMATOID ARTHRITIS
#>   5:                    ANTIVIRAL PROPHYLAXIS
#>  ---                                         
#> 153:                     RHEUMATOID ARTHRITIS
#> 154: ATTENTION DEFICIT/HYPERACTIVITY DISORDER
#> 155:         PERVASIVE DEVELOPMENTAL DISORDER
#> 156:                         FRACTURED SACRUM
#> 157:                         LIPIDS INCREASED
#> 
#> $reac
#>       year quarter primaryid                                   pt     v3
#>      <int>  <char>    <char>                               <char> <lgcl>
#>   1:  2004      q1   4263764             BLOOD PRESSURE INCREASED     NA
#>   2:  2004      q1   4263764 DIABETES MELLITUS INADEQUATE CONTROL     NA
#>   3:  2004      q1   4263927             ACCELERATED HYPERTENSION     NA
#>   4:  2004      q1   4263927                            DIZZINESS     NA
#>   5:  2004      q1   4263927                              FATIGUE     NA
#>  ---                                                                    
#> 401:  2004      q1   4373559                              FATIGUE     NA
#> 402:  2004      q1   4373559                              MYALGIA     NA
#> 403:  2004      q1   4373586   HIGH DENSITY LIPOPROTEIN DECREASED     NA
#> 404:  2004      q1   4373706                     DRUG INTERACTION     NA
#> 405:  2004      q1   4373706             HEPATIC ENZYME INCREASED     NA
#> 
#> $ther
#>       year quarter primaryid dsg_drug_seq start_dt   end_dt   dur dur_cod
#>      <int>  <char>    <char>        <int>    <int>    <int> <int>  <char>
#>   1:  2004      q1   4263764   1004492874 20010101       NA    NA    <NA>
#>   2:  2004      q1   4263764   1004515789 19850101       NA    NA    <NA>
#>   3:  2004      q1   4263764   1004515793 19830101 20010101    NA    <NA>
#>   4:  2004      q1   4263927   1004493661       NA       NA     1     MON
#>   5:  2004      q1   4264001   1004493811 20031202 20031202    NA    <NA>
#>  ---                                                                     
#> 161:  2004      q1   4338901   1004786888 20030205       NA    NA    <NA>
#> 162:  2004      q1   4346899   1004815291 20031101       NA    NA    <NA>
#> 163:  2004      q1   4347265   1004816501 20040128       NA    NA    <NA>
#> 164:  2004      q1   4373559   1004923573 20030923 20031022    NA    <NA>
#> 165:  2004      q1   4373586   1004923625 20031201       NA    NA    <NA>
#>          v7
#>      <lgcl>
#>   1:     NA
#>   2:     NA
#>   3:     NA
#>   4:     NA
#>   5:     NA
#>  ---       
#> 161:     NA
#> 162:     NA
#> 163:     NA
#> 164:     NA
#> 165:     NA
#> 
#> $rpsr
#>       year quarter primaryid rpsr_cod     v3
#>      <int>  <char>    <char>   <char> <lgcl>
#>   1:  2004      q1   4263764      CSM     NA
#>   2:  2004      q1   4263927      FGN     NA
#>   3:  2004      q1   4263927       HP     NA
#>   4:  2004      q1   4263927      LIT     NA
#>   5:  2004      q1   4264001      FGN     NA
#>  ---                                        
#> 103:  2004      q1   4346899      CSM     NA
#> 104:  2004      q1   4347265      CSM     NA
#> 105:  2004      q1   4373559       HP     NA
#> 106:  2004      q1   4373586       HP     NA
#> 107:  2004      q1   4373706       HP     NA
#> 
#> $outc
#>      year quarter primaryid outc_cod     v3
#>     <int>  <char>    <char>   <char> <lgcl>
#>  1:  2004      q1   4263764       HO     NA
#>  2:  2004      q1   4263927       OT     NA
#>  3:  2004      q1   4264001       DE     NA
#>  4:  2004      q1   4264001       HO     NA
#>  5:  2004      q1   4264319       HO     NA
#>  6:  2004      q1   4264319       OT     NA
#>  7:  2004      q1   4266745       HO     NA
#>  8:  2004      q1   4268568       DE     NA
#>  9:  2004      q1   4268568       OT     NA
#> 10:  2004      q1   4268586       OT     NA
#> 11:  2004      q1   4269553       OT     NA
#> 12:  2004      q1   4269590       HO     NA
#> 13:  2004      q1   4269894       HO     NA
#> 14:  2004      q1   4271501       HO     NA
#> 15:  2004      q1   4272687       HO     NA
#> 16:  2004      q1   4274829       HO     NA
#> 17:  2004      q1   4276604       OT     NA
#> 18:  2004      q1   4276994       HO     NA
#> 19:  2004      q1   4277521       OT     NA
#> 20:  2004      q1   4279336       HO     NA
#> 21:  2004      q1   4279336       OT     NA
#> 22:  2004      q1   4280316       HO     NA
#> 23:  2004      q1   4280654       HO     NA
#> 24:  2004      q1   4281199       OT     NA
#> 25:  2004      q1   4281267       RI     NA
#> 26:  2004      q1   4283791       OT     NA
#> 27:  2004      q1   4283960       DE     NA
#> 28:  2004      q1   4283960       DS     NA
#> 29:  2004      q1   4283960       HO     NA
#> 30:  2004      q1   4284513       DE     NA
#> 31:  2004      q1   4284513       HO     NA
#> 32:  2004      q1   4284513       OT     NA
#> 33:  2004      q1   4284876       RI     NA
#> 34:  2004      q1   4285379       DE     NA
#> 35:  2004      q1   4286029       OT     NA
#> 36:  2004      q1   4286712       HO     NA
#> 37:  2004      q1   4286757       HO     NA
#> 38:  2004      q1   4286940       OT     NA
#> 39:  2004      q1   4287019       DE     NA
#> 40:  2004      q1   4290441       HO     NA
#> 41:  2004      q1   4290441       OT     NA
#> 42:  2004      q1   4294153       OT     NA
#> 43:  2004      q1   4294372       DS     NA
#> 44:  2004      q1   4294513       HO     NA
#> 45:  2004      q1   4297303       OT     NA
#> 46:  2004      q1   4297385       DS     NA
#> 47:  2004      q1   4297385       OT     NA
#> 48:  2004      q1   4297385       RI     NA
#> 49:  2004      q1   4299109       HO     NA
#> 50:  2004      q1   4300112       OT     NA
#> 51:  2004      q1   4301549       HO     NA
#> 52:  2004      q1   4301549       RI     NA
#> 53:  2004      q1   4303368       HO     NA
#> 54:  2004      q1   4305272       OT     NA
#> 55:  2004      q1   4306018       HO     NA
#> 56:  2004      q1   4306018       OT     NA
#> 57:  2004      q1   4309138       OT     NA
#> 58:  2004      q1   4309390       OT     NA
#> 59:  2004      q1   4309892       HO     NA
#> 60:  2004      q1   4310779       DE     NA
#> 61:  2004      q1   4310779       HO     NA
#> 62:  2004      q1   4310862       HO     NA
#> 63:  2004      q1   4312635       LT     NA
#> 64:  2004      q1   4314135       OT     NA
#> 65:  2004      q1   4314728       HO     NA
#> 66:  2004      q1   4316846       DE     NA
#> 67:  2004      q1   4316948       HO     NA
#> 68:  2004      q1   4317126       HO     NA
#> 69:  2004      q1   4317723       HO     NA
#> 70:  2004      q1   4318479       HO     NA
#> 71:  2004      q1   4318770       DS     NA
#> 72:  2004      q1   4318770       HO     NA
#> 73:  2004      q1   4319252       LT     NA
#> 74:  2004      q1   4322216       DS     NA
#> 75:  2004      q1   4322216       OT     NA
#> 76:  2004      q1   4322271       DE     NA
#> 77:  2004      q1   4322418       OT     NA
#> 78:  2004      q1   4322977       HO     NA
#> 79:  2004      q1   4322977       LT     NA
#> 80:  2004      q1   4322977       OT     NA
#> 81:  2004      q1   4322977       RI     NA
#> 82:  2004      q1   4324746       DS     NA
#> 83:  2004      q1   4324746       HO     NA
#> 84:  2004      q1   4324746       OT     NA
#> 85:  2004      q1   4325623       DS     NA
#> 86:  2004      q1   4325623       HO     NA
#> 87:  2004      q1   4325623       OT     NA
#> 88:  2004      q1   4326344       HO     NA
#> 89:  2004      q1   4326344       LT     NA
#> 90:  2004      q1   4326344       RI     NA
#> 91:  2004      q1   4327943       LT     NA
#> 92:  2004      q1   4330031       LT     NA
#> 93:  2004      q1   4331457       OT     NA
#> 94:  2004      q1   4373559       RI     NA
#> 95:  2004      q1   4373586       OT     NA
#> 96:  2004      q1   4373706       OT     NA
#>      year quarter primaryid outc_cod     v3
#> 
faers_year(data)
#> [1] 2004
faers_quarter(data)
#> [1] "q1"
faers_period(data)
#>     year quarter
#>    <int>  <char>
#> 1:  2004      q1
faers_meddra(data)
#> NULL
faers_deleted_cases(data)
#> character(0)