MedDRA class

Provide a container for MedDRA Data file

meddra(path, add_smq = FALSE, primary_soc = FALSE)

# S4 method for class 'MedDRA'
show(object)

meddra_hierarchy(object, ...)

# S4 method for class 'MedDRA'
meddra_hierarchy(object)

meddra_smq(object, ...)

# S4 method for class 'MedDRA'
meddra_smq(object)

meddra_version(object, ...)

# S4 method for class 'MedDRA'
meddra_version(object)

Arguments

path

A string, define the path of MedDRA directory.

add_smq

A bool, indicates whether Standardised MedDRA Queries (SMQ) should be added. If TRUE, "smq_content.asc", and "smq_list.asc" must exist.

primary_soc

A bool, indicates whether keep primary soc only.

object

A MedDRA object.

...

Other arguments passed to specific methods.

Value

meddra: A MedDRA object.

meddra_hierarchy: Extract the hierarchy slot.

meddra_smq: Extract the smq slot.

meddra_version: Extract the version slot.

Slots

hierarchy

A data.table or NULL representing the meddra hierarchy data. There are five levels to the MedDRA hierarchy, arranged from very specific to very general.

smq

A data.table or NULL representing the meddra smq data. Standardised MedDRA Queries (SMQs) are used to support signal detection and monitoring. SMQs are validated, standard sets of MedDRA terms. These sets of terms have undergone extensive review, testing, analysis and expert discussion. SMQs represent a variety of safety topics of regulatory interest (e.g., SMQ Severe cutaneous adverse reactions, SMQ Anaphylactic reaction).

version

A string, the version of MedDRA.

See also