Merge all FAERS field data into one

Merge all FAERS field data into one

faers_merge(object, ...)

# S4 method for class 'FAERSascii'
faers_merge(object, fields = NULL, all = TRUE, all.x = all, all.y = all)

# S4 method for class 'FAERSxml'
faers_merge(object)

# S4 method for class 'ANY'
faers_merge(object)

Arguments

object

A FAERSascii or FAERSxml object.

...

Other arguments passed to specific methods.

fields

A character vector specifying the fields to use. Only values "demo", "drug", "indi", "ther", "reac", "rpsr", and "outc" can be used.

all

logical; all = TRUE is shorthand to save setting both all.x = TRUE and all.y = TRUE.

all.x

logical; if TRUE, rows from x which have no matching row in y are included. These rows will have 'NA's in the columns that are usually filled with values from y. The default is FALSE so that only rows with data from both x and y are included in the output.

all.y

logical; analogous to all.x above.

Value

A data.table object.

Details

Each pair of field data are merged based on "year", "quarter" and "primaryid". In cases where any pair of data contains information related to "drug_seq" elements, such as "drug_seq", "indi_drug_seq", or "dsg_drug_seq", "drug_seq" will be aligned as well. fields shall be organized in the subsequent sequence: 'demo', 'drug', 'indi', 'reac', 'ther', 'rpsr', and 'outc' and the merging sequence will correspondingly adhere to this order. Only the initial instance, of the "caseid" column will be preserved.

Note

You'd better only merge necessary data, otherwise a lot of memory will be consumed to merge all fields data.

Examples

# you must change `dir`, as the file included in the package is sampled
data <- faers(2004, "q1",
    dir = system.file("extdata", package = "faers"),
    compress_dir = tempdir()
)
#> Finding 1 file already downloaded: aers_ascii_2004q1.zip
faers_merge(data, "indi") # only one field is just like faers_get()
#>       year quarter primaryid indi_drug_seq
#>      <int>  <char>    <char>         <int>
#>   1:  2004      q1   4263927    1004493661
#>   2:  2004      q1   4264001    1004493811
#>   3:  2004      q1   4264001    1004520441
#>   4:  2004      q1   4264001    1004520538
#>   5:  2004      q1   4264319    1004494389
#>  ---                                      
#> 153:  2004      q1   4338901    1004786888
#> 154:  2004      q1   4346899    1004815291
#> 155:  2004      q1   4346899    1004815291
#> 156:  2004      q1   4347265    1004816501
#> 157:  2004      q1   4373559    1004923573
#>                                       indi_pt
#>                                        <char>
#>   1:  DIABETES MELLITUS NON-INSULIN-DEPENDENT
#>   2:                     RHEUMATOID ARTHRITIS
#>   3:                     RHEUMATOID ARTHRITIS
#>   4:                     RHEUMATOID ARTHRITIS
#>   5:                    ANTIVIRAL PROPHYLAXIS
#>  ---                                         
#> 153:                     RHEUMATOID ARTHRITIS
#> 154: ATTENTION DEFICIT/HYPERACTIVITY DISORDER
#> 155:         PERVASIVE DEVELOPMENTAL DISORDER
#> 156:                         FRACTURED SACRUM
#> 157:                         LIPIDS INCREASED
faers_merge(data, c("demo", "indi"))
#>       year quarter primaryid  caseid i_f_code foll_seq     image event_dt
#>      <int>  <char>    <char>  <char>   <char>    <int>    <char>    <int>
#>   1:  2004      q1   4263764 4060920        I       NA 4263764-6 20020101
#>   2:  2004      q1   4263927 4064250        I       NA 4263927-X       NA
#>   3:  2004      q1   4264001 4062524        I       NA 4264001-9 20031218
#>   4:  2004      q1   4264001 4062524        I       NA 4264001-9 20031218
#>   5:  2004      q1   4264001 4062524        I       NA 4264001-9 20031218
#>  ---                                                                     
#> 167:  2004      q1   4346899 4135105        I       NA 4346899-9 20030101
#> 168:  2004      q1   4347265 4133168        I       NA 4347265-2 20040128
#> 169:  2004      q1   4373559 4151463        F        1 4373559-0 20030923
#> 170:  2004      q1   4373586 4111031        I       NA 4373586-3 20040123
#> 171:  2004      q1   4373706 4151560        I       NA 4373706-0       NA
#>        mfr_dt   fda_dt rept_cod               mfr_num
#>         <int>    <int>   <char>                <char>
#>   1: 20031219 20040102      EXP          USA031255171
#>   2: 20031209 20040102      EXP             B0317710A
#>   3: 20031219 20040102      EXP JP-JNJFOC-20031204393
#>   4: 20031219 20040102      EXP JP-JNJFOC-20031204393
#>   5: 20031219 20040102      EXP JP-JNJFOC-20031204393
#>  ---                                                 
#> 167: 20031222 20040329      PER          USA031255274
#> 168: 20040217 20040325      PER          USA040158011
#> 169: 20031209 20040305      PER           2003UW14133
#> 170: 20040123 20040305      PER           2004UW01678
#> 171: 20040106 20040305      PER           2004UW00207
#>                                    mfr_sndr   age age_cod gender  e_sub    wt
#>                                      <char> <num>  <char> <char> <char> <num>
#>   1:                  ELI LILLY AND COMPANY    68      YR      F      N  82.0
#>   2: GLAXOSMITHKLINE GLOBAL CLINICAL SAFETY    58      YR      F      N    NA
#>   3:                         CENTOCOR, INC.    53      YR      F      N  36.8
#>   4:                         CENTOCOR, INC.    53      YR      F      N  36.8
#>   5:                         CENTOCOR, INC.    53      YR      F      N  36.8
#>  ---                                                                         
#> 167:                  ELI LILLY AND COMPANY     6      YR      M      N  27.0
#> 168:                  ELI LILLY AND COMPANY    76      YR      F      N    NA
#> 169:            ASTRAZENECA PHARMACEUTICALS    78      YR      M      N 243.0
#> 170:            ASTRAZENECA PHARMACEUTICALS    56      YR      M      N    NA
#> 171:            ASTRAZENECA PHARMACEUTICALS    NA    <NA>   <NA>      N    NA
#>      wt_cod  rept_dt occp_cod death_dt to_mfr confid    v23 caseversion
#>      <char>    <int>   <char>   <lgcl> <char> <char> <lgcl>       <int>
#>   1:     KG 20031223     <NA>       NA   <NA>   <NA>     NA           0
#>   2:   <NA> 20031219     <NA>       NA   <NA>   <NA>     NA           0
#>   3:     KG 20031231       MD       NA   <NA>   <NA>     NA           0
#>   4:     KG 20031231       MD       NA   <NA>   <NA>     NA           0
#>   5:     KG 20031231       MD       NA   <NA>   <NA>     NA           0
#>  ---                                                                   
#> 167:     KG 20040301     <NA>       NA   <NA>   <NA>     NA           0
#> 168:   <NA> 20040309     <NA>       NA   <NA>   <NA>     NA           0
#> 169:    LBS 20040223       MD       NA   <NA>   <NA>     NA           0
#> 170:   <NA> 20040223       MD       NA   <NA>   <NA>     NA           0
#> 171:   <NA> 20040223     <NA>       NA   <NA>   <NA>     NA           0
#>      age_in_years country_code    sex indi_drug_seq
#>             <num>       <char> <char>         <int>
#>   1:           68         <NA> Female            NA
#>   2:           58         <NA> Female    1004493661
#>   3:           53         <NA> Female    1004493811
#>   4:           53         <NA> Female    1004520441
#>   5:           53         <NA> Female    1004520538
#>  ---                                               
#> 167:            6         <NA>   Male    1004815291
#> 168:           76         <NA> Female    1004816501
#> 169:           78         <NA>   Male    1004923573
#> 170:           56         <NA>   Male            NA
#> 171:           NA         <NA>   <NA>            NA
#>                                      indi_pt
#>                                       <char>
#>   1:                                    <NA>
#>   2: DIABETES MELLITUS NON-INSULIN-DEPENDENT
#>   3:                    RHEUMATOID ARTHRITIS
#>   4:                    RHEUMATOID ARTHRITIS
#>   5:                    RHEUMATOID ARTHRITIS
#>  ---                                        
#> 167:        PERVASIVE DEVELOPMENTAL DISORDER
#> 168:                        FRACTURED SACRUM
#> 169:                        LIPIDS INCREASED
#> 170:                                    <NA>
#> 171:                                    <NA>