List of FAERS data

The function lists the metadata for the FAERS databases currently available to download.

faers_meta(force = FALSE, internal = !curl::has_internet())

Arguments

force

A boolean value. If set to TRUE, it indicates the retrieval of information about all records' metadata in the FAERS Quarterly Data Extract Files Site, bypassing the cache.

internal

A boolean value. It determines whether to use the internal data associated with the package.

Value

A data.table reporting years, period, quarter, and file urls and file sizes.

See also

Examples

faers_meta()
#> → Reading html:
#>   <https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html>
#> → Writing FAERS metadata into cache file
#>   /home/wzy/.cache/R/faers/faers/metadata/faers_meta_data.rds
#>      year quarter             period
#>     <int>  <char>             <char>
#>  1:  2025      q4 October - December
#>  2:  2025      q3   July - September
#>  3:  2025      q2       April - June
#>  4:  2025      q1    January - March
#>  5:  2024      q4 October - December
#>  6:  2024      q3   July - September
#>  7:  2024      q2       April - June
#>  8:  2024      q1    January - March
#>  9:  2023      q4 October - December
#> 10:  2023      q3   July - September
#> 11:  2023      q2       April - June
#> 12:  2023      q1    January - March
#> 13:  2022      q4 October - December
#> 14:  2022      q3   July - September
#> 15:  2022      q2       April - June
#> 16:  2022      q1    January - March
#> 17:  2021      q4 October - December
#> 18:  2021      q3   July - September
#> 19:  2021      q2       April - June
#> 20:  2021      q1    January - March
#> 21:  2020      q4 October - December
#> 22:  2020      q3   July - September
#> 23:  2020      q2       April - June
#> 24:  2020      q1    January - March
#> 25:  2019      q4 October - December
#> 26:  2019      q3   July - September
#> 27:  2019      q2       April - June
#> 28:  2019      q1    January - March
#> 29:  2018      q4 October - December
#> 30:  2018      q3   July - September
#> 31:  2018      q2       April - June
#> 32:  2018      q1    January - March
#> 33:  2017      q4 October - December
#> 34:  2017      q3   July - September
#> 35:  2017      q2       April - June
#> 36:  2017      q1    January - March
#> 37:  2016      q4 October - December
#> 38:  2016      q3   July - September
#> 39:  2016      q2       April - June
#> 40:  2016      q1    January - March
#> 41:  2015      q4 October - December
#> 42:  2015      q3   July - September
#> 43:  2015      q2       April - June
#> 44:  2015      q1    January - March
#> 45:  2014      q4 October - December
#> 46:  2014      q3   July - September
#> 47:  2014      q2       April - June
#> 48:  2014      q1    January - March
#> 49:  2013      q4 October - December
#> 50:  2013      q3   July - September
#> 51:  2013      q2       April - June
#> 52:  2013      q1    January - March
#> 53:  2012      q4 October - December
#> 54:  2012      q3   July - September
#> 55:  2012      q2       April - June
#> 56:  2012      q1    January - March
#> 57:  2011      q4 October - December
#> 58:  2011      q3   July - September
#> 59:  2011      q2       April - June
#> 60:  2011      q1    January - March
#> 61:  2010      q4 October - December
#> 62:  2010      q3   July - September
#> 63:  2010      q2       April - June
#> 64:  2010      q1    January - March
#> 65:  2009      q4 October - December
#> 66:  2009      q3   July - September
#> 67:  2009      q2       April - June
#> 68:  2009      q1    January - March
#> 69:  2008      q4 October - December
#> 70:  2008      q3   July - September
#> 71:  2008      q2       April - June
#> 72:  2008      q1    January - March
#> 73:  2007      q4 October - December
#> 74:  2007      q3   July - September
#> 75:  2007      q2       April - June
#> 76:  2007      q1    January - March
#> 77:  2006      q4 October - December
#> 78:  2006      q3   July - September
#> 79:  2006      q2       April - June
#> 80:  2006      q1    January - March
#> 81:  2005      q4 October - December
#> 82:  2005      q3   July - September
#> 83:  2005      q2       April - June
#> 84:  2005      q1    January - March
#> 85:  2004      q4 October - December
#> 86:  2004      q3   July - September
#> 87:  2004      q2       April - June
#> 88:  2004      q1    January - March
#>      year quarter             period
#>                                                     ascii_urls ascii_file_size
#>                                                         <char>          <char>
#>  1: https://fis.fda.gov/content/Exports/faers_ascii_2025Q4.zip            64MB
#>  2: https://fis.fda.gov/content/Exports/faers_ascii_2025q3.zip          73.1MB
#>  3: https://fis.fda.gov/content/Exports/faers_ascii_2025q2.zip          61.5MB
#>  4: https://fis.fda.gov/content/Exports/faers_ascii_2025q1.zip          64.3MB
#>  5: https://fis.fda.gov/content/Exports/faers_ascii_2024Q4.zip          65.6MB
#>  6: https://fis.fda.gov/content/Exports/faers_ascii_2024q3.zip          63.6MB
#>  7: https://fis.fda.gov/content/Exports/faers_ascii_2024q2.zip          63.9MB
#>  8: https://fis.fda.gov/content/Exports/faers_ascii_2024q1.zip          63.1MB
#>  9: https://fis.fda.gov/content/Exports/faers_ascii_2023Q4.zip          69.1MB
#> 10: https://fis.fda.gov/content/Exports/faers_ascii_2023Q3.zip          60.1MB
#> 11: https://fis.fda.gov/content/Exports/faers_ascii_2023q2.zip          64.5MB
#> 12: https://fis.fda.gov/content/Exports/faers_ascii_2023q1.zip          64.3MB
#> 13: https://fis.fda.gov/content/Exports/faers_ascii_2022Q4.zip            69MB
#> 14: https://fis.fda.gov/content/Exports/faers_ascii_2022Q3.zip          63.2MB
#> 15: https://fis.fda.gov/content/Exports/faers_ascii_2022q2.zip            63MB
#> 16: https://fis.fda.gov/content/Exports/faers_ascii_2022q1.zip          64.7MB
#> 17: https://fis.fda.gov/content/Exports/faers_ascii_2021Q4.zip            59MB
#> 18: https://fis.fda.gov/content/Exports/faers_ascii_2021Q3.zip            70MB
#> 19: https://fis.fda.gov/content/Exports/faers_ascii_2021Q2.zip            66MB
#> 20: https://fis.fda.gov/content/Exports/faers_ascii_2021Q1.zip            69MB
#> 21: https://fis.fda.gov/content/Exports/faers_ascii_2020Q4.zip            71MB
#> 22: https://fis.fda.gov/content/Exports/faers_ascii_2020Q3.zip            64MB
#> 23: https://fis.fda.gov/content/Exports/faers_ascii_2020Q2.zip            66MB
#> 24: https://fis.fda.gov/content/Exports/faers_ascii_2020Q1.zip            65MB
#> 25: https://fis.fda.gov/content/Exports/faers_ascii_2019Q4.zip            60MB
#> 26: https://fis.fda.gov/content/Exports/faers_ascii_2019Q3.zip            62MB
#> 27: https://fis.fda.gov/content/Exports/faers_ascii_2019Q2.zip            62MB
#> 28: https://fis.fda.gov/content/Exports/faers_ascii_2019Q1.zip            56MB
#> 29: https://fis.fda.gov/content/Exports/faers_ascii_2018q4.zip            60MB
#> 30: https://fis.fda.gov/content/Exports/faers_ascii_2018q3.zip            60MB
#> 31: https://fis.fda.gov/content/Exports/faers_ascii_2018q2.zip            60MB
#> 32: https://fis.fda.gov/content/Exports/faers_ascii_2018q1.zip            52MB
#> 33: https://fis.fda.gov/content/Exports/faers_ascii_2017q4.zip            41MB
#> 34: https://fis.fda.gov/content/Exports/faers_ascii_2017q3.zip            48MB
#> 35: https://fis.fda.gov/content/Exports/faers_ascii_2017q2.zip            46MB
#> 36: https://fis.fda.gov/content/Exports/faers_ascii_2017q1.zip            48MB
#> 37: https://fis.fda.gov/content/Exports/faers_ascii_2016q4.zip            44MB
#> 38: https://fis.fda.gov/content/Exports/faers_ascii_2016q3.zip            46MB
#> 39: https://fis.fda.gov/content/Exports/faers_ascii_2016q2.zip            44MB
#> 40: https://fis.fda.gov/content/Exports/faers_ascii_2016q1.zip            46MB
#> 41: https://fis.fda.gov/content/Exports/faers_ascii_2015q4.zip            42MB
#> 42: https://fis.fda.gov/content/Exports/faers_ascii_2015q3.zip            47MB
#> 43: https://fis.fda.gov/content/Exports/faers_ascii_2015q2.zip            38MB
#> 44: https://fis.fda.gov/content/Exports/faers_ascii_2015q1.zip            39MB
#> 45: https://fis.fda.gov/content/Exports/faers_ascii_2014q4.zip            28MB
#> 46: https://fis.fda.gov/content/Exports/faers_ascii_2014q3.zip            28MB
#> 47: https://fis.fda.gov/content/Exports/faers_ascii_2014q2.zip            25MB
#> 48: https://fis.fda.gov/content/Exports/faers_ascii_2014q1.zip            30MB
#> 49: https://fis.fda.gov/content/Exports/faers_ascii_2013q4.zip            26MB
#> 50: https://fis.fda.gov/content/Exports/faers_ascii_2013q3.zip            22MB
#> 51: https://fis.fda.gov/content/Exports/faers_ascii_2013q2.zip            21MB
#> 52: https://fis.fda.gov/content/Exports/faers_ascii_2013q1.zip            25MB
#> 53: https://fis.fda.gov/content/Exports/faers_ascii_2012q4.zip            28MB
#> 54:  https://fis.fda.gov/content/Exports/aers_ascii_2012q3.zip            16MB
#> 55:  https://fis.fda.gov/content/Exports/aers_ascii_2012q2.zip            25MB
#> 56:  https://fis.fda.gov/content/Exports/aers_ascii_2012q1.zip            26MB
#> 57:  https://fis.fda.gov/content/Exports/aers_ascii_2011q4.zip            23MB
#> 58:  https://fis.fda.gov/content/Exports/aers_ascii_2011q3.zip            23MB
#> 59:  https://fis.fda.gov/content/Exports/aers_ascii_2011q2.zip            23MB
#> 60:  https://fis.fda.gov/content/Exports/aers_ascii_2011q1.zip            21MB
#> 61:  https://fis.fda.gov/content/Exports/aers_ascii_2010q4.zip            20MB
#> 62:  https://fis.fda.gov/content/Exports/aers_ascii_2010q3.zip            22MB
#> 63:  https://fis.fda.gov/content/Exports/aers_ascii_2010q2.zip            17MB
#> 64:  https://fis.fda.gov/content/Exports/aers_ascii_2010q1.zip            16MB
#> 65:  https://fis.fda.gov/content/Exports/aers_ascii_2009q4.zip            16MB
#> 66:  https://fis.fda.gov/content/Exports/aers_ascii_2009q3.zip            16MB
#> 67:  https://fis.fda.gov/content/Exports/aers_ascii_2009q2.zip            14MB
#> 68:  https://fis.fda.gov/content/Exports/aers_ascii_2009q1.zip            13MB
#> 69:  https://fis.fda.gov/content/Exports/aers_ascii_2008q4.zip            13MB
#> 70:  https://fis.fda.gov/content/Exports/aers_ascii_2008q3.zip            13MB
#> 71:  https://fis.fda.gov/content/Exports/aers_ascii_2008q2.zip            12MB
#> 72:  https://fis.fda.gov/content/Exports/aers_ascii_2008q1.zip            12MB
#> 73:  https://fis.fda.gov/content/Exports/aers_ascii_2007q4.zip            12MB
#> 74:  https://fis.fda.gov/content/Exports/aers_ascii_2007q3.zip           9.9MB
#> 75:  https://fis.fda.gov/content/Exports/aers_ascii_2007q2.zip           9.5MB
#> 76:  https://fis.fda.gov/content/Exports/aers_ascii_2007q1.zip           9.6MB
#> 77:  https://fis.fda.gov/content/Exports/aers_ascii_2006q4.zip           9.1MB
#> 78:  https://fis.fda.gov/content/Exports/aers_ascii_2006q3.zip           8.5MB
#> 79:  https://fis.fda.gov/content/Exports/aers_ascii_2006q2.zip           9.7MB
#> 80:  https://fis.fda.gov/content/Exports/aers_ascii_2006q1.zip            11MB
#> 81:  https://fis.fda.gov/content/Exports/aers_ascii_2005q4.zip           9.7MB
#> 82:  https://fis.fda.gov/content/Exports/aers_ascii_2005q3.zip           9.3MB
#> 83:  https://fis.fda.gov/content/Exports/aers_ascii_2005q2.zip           9.5MB
#> 84:  https://fis.fda.gov/content/Exports/aers_ascii_2005q1.zip           8.5MB
#> 85:  https://fis.fda.gov/content/Exports/aers_ascii_2004q4.zip           8.1MB
#> 86:  https://fis.fda.gov/content/Exports/aers_ascii_2004q3.zip           8.2MB
#> 87:  https://fis.fda.gov/content/Exports/aers_ascii_2004q2.zip           7.1MB
#> 88:  https://fis.fda.gov/content/Exports/aers_ascii_2004q1.zip           7.3MB
#>                                                     ascii_urls ascii_file_size
#>                                                     xml_urls xml_file_size
#>                                                       <char>        <char>
#>  1: https://fis.fda.gov/content/Exports/faers_xml_2025Q4.zip         106MB
#>  2: https://fis.fda.gov/content/Exports/faers_xml_2025q3.zip         115MB
#>  3: https://fis.fda.gov/content/Exports/faers_xml_2025q2.zip         107MB
#>  4: https://fis.fda.gov/content/Exports/faers_xml_2025q1.zip         111MB
#>  5: https://fis.fda.gov/content/Exports/faers_xml_2024Q4.zip         118MB
#>  6: https://fis.fda.gov/content/Exports/faers_xml_2024q3.zip         129MB
#>  7: https://fis.fda.gov/content/Exports/faers_xml_2024q2.zip         126MB
#>  8: https://fis.fda.gov/content/Exports/faers_xml_2024q1.zip         128MB
#>  9: https://fis.fda.gov/content/Exports/faers_xml_2023Q4.zip         140MB
#> 10: https://fis.fda.gov/content/Exports/faers_xml_2023Q3.zip         123MB
#> 11: https://fis.fda.gov/content/Exports/faers_xml_2023q2.zip         130MB
#> 12: https://fis.fda.gov/content/Exports/faers_xml_2023q1.zip         133MB
#> 13: https://fis.fda.gov/content/Exports/faers_xml_2022Q4.zip         144MB
#> 14: https://fis.fda.gov/content/Exports/faers_xml_2022Q3.zip         132MB
#> 15: https://fis.fda.gov/content/Exports/faers_xml_2022q2.zip         140MB
#> 16: https://fis.fda.gov/content/Exports/faers_xml_2022q1.zip         136MB
#> 17: https://fis.fda.gov/content/Exports/faers_xml_2021Q4.zip         123MB
#> 18: https://fis.fda.gov/content/Exports/faers_xml_2021Q3.zip         132MB
#> 19: https://fis.fda.gov/content/Exports/faers_xml_2021Q2.zip         123MB
#> 20: https://fis.fda.gov/content/Exports/faers_xml_2021Q1.zip         130MB
#> 21: https://fis.fda.gov/content/Exports/faers_xml_2020Q4.zip         131MB
#> 22: https://fis.fda.gov/content/Exports/faers_xml_2020Q3.zip         121MB
#> 23: https://fis.fda.gov/content/Exports/faers_xml_2020Q2.zip         123MB
#> 24: https://fis.fda.gov/content/Exports/faers_xml_2020Q1.zip         125MB
#> 25: https://fis.fda.gov/content/Exports/faers_xml_2019Q4.zip         113MB
#> 26: https://fis.fda.gov/content/Exports/faers_xml_2019Q3.zip         118MB
#> 27: https://fis.fda.gov/content/Exports/faers_xml_2019Q2.zip         118MB
#> 28: https://fis.fda.gov/content/Exports/faers_xml_2019Q1.zip         103MB
#> 29: https://fis.fda.gov/content/Exports/faers_xml_2018q4.zip         112MB
#> 30: https://fis.fda.gov/content/Exports/faers_xml_2018q3.zip         112MB
#> 31: https://fis.fda.gov/content/Exports/faers_xml_2018q2.zip         112MB
#> 32: https://fis.fda.gov/content/Exports/faers_xml_2018q1.zip          94MB
#> 33: https://fis.fda.gov/content/Exports/faers_xml_2017q4.zip          76MB
#> 34: https://fis.fda.gov/content/Exports/faers_xml_2017q3.zip          91MB
#> 35: https://fis.fda.gov/content/Exports/faers_xml_2017q2.zip          86MB
#> 36: https://fis.fda.gov/content/Exports/faers_xml_2017q1.zip          91MB
#> 37: https://fis.fda.gov/content/Exports/faers_xml_2016q4.zip          82MB
#> 38: https://fis.fda.gov/content/Exports/faers_xml_2016q3.zip          87MB
#> 39: https://fis.fda.gov/content/Exports/faers_xml_2016q2.zip          81MB
#> 40: https://fis.fda.gov/content/Exports/faers_xml_2016q1.zip          84MB
#> 41: https://fis.fda.gov/content/Exports/faers_xml_2015q4.zip          77MB
#> 42: https://fis.fda.gov/content/Exports/faers_xml_2015q3.zip          88MB
#> 43: https://fis.fda.gov/content/Exports/faers_xml_2015q2.zip          70MB
#> 44: https://fis.fda.gov/content/Exports/faers_xml_2015q1.zip          72MB
#> 45: https://fis.fda.gov/content/Exports/faers_xml_2014q4.zip          53MB
#> 46: https://fis.fda.gov/content/Exports/faers_xml_2014q3.zip          54MB
#> 47: https://fis.fda.gov/content/Exports/faers_xml_2014q2.zip          44MB
#> 48: https://fis.fda.gov/content/Exports/faers_xml_2014q1.zip          52MB
#> 49: https://fis.fda.gov/content/Exports/faers_xml_2013q4.zip          46MB
#> 50: https://fis.fda.gov/content/Exports/faers_xml_2013q3.zip          40MB
#> 51: https://fis.fda.gov/content/Exports/faers_xml_2013q2.zip          38MB
#> 52: https://fis.fda.gov/content/Exports/faers_xml_2013q1.zip          44MB
#> 53: https://fis.fda.gov/content/Exports/faers_xml_2012q4.zip          50MB
#> 54: https://fis.fda.gov/content/Exports/aers_sgml_2012q3.zip          21MB
#> 55: https://fis.fda.gov/content/Exports/aers_sgml_2012q2.zip          32MB
#> 56: https://fis.fda.gov/content/Exports/aers_sgml_2012q1.zip          33MB
#> 57: https://fis.fda.gov/content/Exports/aers_sgml_2011q4.zip          29MB
#> 58: https://fis.fda.gov/content/Exports/aers_sgml_2011q3.zip          29MB
#> 59: https://fis.fda.gov/content/Exports/aers_sgml_2011q2.zip          29MB
#> 60: https://fis.fda.gov/content/Exports/aers_sgml_2011q1.zip          26MB
#> 61: https://fis.fda.gov/content/Exports/aers_sgml_2010q4.zip          25MB
#> 62: https://fis.fda.gov/content/Exports/aers_sgml_2010q3.zip          28MB
#> 63: https://fis.fda.gov/content/Exports/aers_sgml_2010q2.zip          22MB
#> 64: https://fis.fda.gov/content/Exports/aers_sgml_2010q1.zip          20MB
#> 65: https://fis.fda.gov/content/Exports/aers_sgml_2009q4.zip          20MB
#> 66: https://fis.fda.gov/content/Exports/aers_sgml_2009q3.zip          19MB
#> 67: https://fis.fda.gov/content/Exports/aers_sgml_2009q2.zip          18MB
#> 68: https://fis.fda.gov/content/Exports/aers_sgml_2009q1.zip          16MB
#> 69: https://fis.fda.gov/content/Exports/aers_sgml_2008q4.zip          16MB
#> 70: https://fis.fda.gov/content/Exports/aers_sgml_2008q3.zip          16MB
#> 71: https://fis.fda.gov/content/Exports/aers_sgml_2008q2.zip          16MB
#> 72: https://fis.fda.gov/content/Exports/aers_sgml_2008q1.zip          15MB
#> 73: https://fis.fda.gov/content/Exports/aers_sgml_2007q4.zip          14MB
#> 74: https://fis.fda.gov/content/Exports/aers_sgml_2007q3.zip          13MB
#> 75: https://fis.fda.gov/content/Exports/aers_sgml_2007q2.zip          12MB
#> 76: https://fis.fda.gov/content/Exports/aers_sgml_2007q1.zip          12MB
#> 77: https://fis.fda.gov/content/Exports/aers_sgml_2006q4.zip          12MB
#> 78: https://fis.fda.gov/content/Exports/aers_sgml_2006q3.zip          11MB
#> 79: https://fis.fda.gov/content/Exports/aers_sgml_2006q2.zip          13MB
#> 80: https://fis.fda.gov/content/Exports/aers_sgml_2006q1.zip          13MB
#> 81: https://fis.fda.gov/content/Exports/aers_sgml_2005q4.zip          12MB
#> 82: https://fis.fda.gov/content/Exports/aers_sgml_2005q3.zip          12MB
#> 83: https://fis.fda.gov/content/Exports/aers_sgml_2005q2.zip          12MB
#> 84: https://fis.fda.gov/content/Exports/aers_sgml_2005q1.zip          11MB
#> 85: https://fis.fda.gov/content/Exports/aers_sgml_2004q4.zip         9.8MB
#> 86: https://fis.fda.gov/content/Exports/aers_sgml_2004q3.zip         9.7MB
#> 87: https://fis.fda.gov/content/Exports/aers_sgml_2004q2.zip         8.6MB
#> 88: https://fis.fda.gov/content/Exports/aers_sgml_2004q1.zip         8.8MB
#>                                                     xml_urls xml_file_size